+++ resources/views/components/layouts/app.blade.php
@@ -22,10 +22,7 @@
     <meta name="csrf-token" content="{{ csrf_token() }}">
     <title>{{ $title ? $title . ' - ' . config('app.name') : config('app.name') }}</title>
 
-    <link
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&display=block"
-        rel="stylesheet">
-
+    @fonts
     @vite(['resources/css/app.css', 'resources/js/app.js'])
     @livewireStyles
 </head>