+++ resources/views/components/layouts/auth.blade.php
@@ -13,10 +13,7 @@
     <meta name="description" content="{{ $description }}">
     <title>{{ $title ? $title . ' | ' . config('app.name') : config('app.name') }}</title>
 
-    <link
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&display=block"
-        rel="stylesheet">
-
+    @fonts
     @vite(['resources/css/app.css', 'resources/js/app.js'])
     @livewireStyles
 </head>